Add startup listener to enable environment checks before starting.
For now only added the check if a Spring property contains a valid
message, but this can be used for more checks at well at different
stages.

This way users of the application don't get a weird stacktrace from
Spring we can really help the users point at the the missing
expected configuration.
Nothing is more annoying about an application starting with huge
stacktrace when all you miss is a environment property.