fix: huge performance issue with Thymeleaf
The loop in header.html to generate all links to all lessons in the dropdown is a performance hit.

Traced it back to `@{/challenge/}` trying to build a link in `StandardLinkBuilder` which results in Spring resolving an url path, resulting in eventually calling CachingResourceResolver#resolveUrlPathInternal which returns null as `@{challenge/}` is not a valid url.

This a lot of time, as far as I can tell the syntax in Thymeleaf for building a link did not change. Since we already have the link in `ChallengeUI` we can simply avoid the concatenation and `@{/challenge/}` usage.

This solves the performance problem at hand, results go from 32 request/sec back to around 445 request/sec.